Completed the Implementation of Clear task and Used lottie to improve UI

+++ app/src/main/java/com/example/task/CompletedFragment.kt
@@ -1,5 +1,6 @@
 package com.example.task
 
+import android.os.Build
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
@@ -7,6 +8,8 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.PopupMenu
 import android.widget.Toast
+import androidx.annotation.RequiresApi
+import androidx.appcompat.app.AlertDialog
 import androidx.fragment.app.activityViewModels
 import androidx.lifecycle.lifecycleScope
 import androidx.navigation.fragment.findNavController
@@ -43,8 +46,22 @@ class CompletedFragment : Fragment() {
 
     }
 
+    @RequiresApi(Build.VERSION_CODES.M)
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
+        val lottieAnimationView = binding.lottieCompleted
+        sharedViewModel.completedTask.observe(viewLifecycleOwner) {
+            if (it.isEmpty()){
+                lottieAnimationView.visibility = View.VISIBLE
+                lottieAnimationView.setAnimation(R.raw.completed_animation)
+                lottieAnimationView.loop(true)
+                lottieAnimationView.playAnimation()
+            }
+            else{
+                lottieAnimationView.visibility = View.GONE
+            }
+        }
+
         binding.recyclerViewForFragmentCompleted.layoutManager = LinearLayoutManager(activity)
         lifecycleScope.launch{
             sharedViewModel.completedTask.observe(viewLifecycleOwner) {
@@ -57,7 +74,15 @@ class CompletedFragment : Fragment() {
             }
         }
         binding.clearAll.setOnClickListener {
-            sharedViewModel.deleteAllFromCompleted()
+            val alertDialog = AlertDialog.Builder(requireContext())
+            alertDialog.setMessage(R.string.delete_all_from_completed)
+            alertDialog.setPositiveButton(android.R.string.yes){ _, _ ->
+                sharedViewModel.deleteAllFromCompleted()
+            }
+            alertDialog.setNegativeButton(android.R.string.no){ dialog,_ ->
+                dialog.cancel()
+            }
+            alertDialog.show()
         }
     }
